src\pygubu\component\plugin_engine.py unnecessary-pass
The Abstract methods configure_for_preview and  ensure_visibility_in_preview of the interface class IDesignerPluginare left as pass. 

Therefore, they might be accidentally used, without implementation. 
Raising NotImplementedError to avoid it.

However, other methods in the same file are decorated with abstractmethod.
Should discuss with owners if it is intended to have no action as default.